Separated concerns related to guidance cards into its own file. Placed guidance cards under its own action sheet selection.
